requirements.txt
Babel==2.5.0
Django==1.8.2
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.9.0
PyNaCl==1.1.2
PyYAML==3.12
Pygments==2.2.0
Sphinx==1.6.3
XlsxWriter==0.7.7
alabaster==0.7.10
ansible==2.2.0.0
appnope==0.1.0
asn1crypto==0.22.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
baidu-aip==1.6.1.0
bcrypt==3.1.4
beautifulsoup4==4.4.0
bleach==2.0.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
configparser==3.5.0
construct==2.5.3
cryptography==2.1.1
decorator==4.1.2
diff-match-patch==20121119
django-bootstrap-form==3.2
django-ckeditor==5.3.0
django-classy-tags==0.8.0
django-cms==3.1.3
django-contrib-comments==1.6.1
django-cors-headers==1.1.0
django-crispy-forms==1.7.0
django-extensions==1.7.4
django-formtools==2.1
django-import-export==0.5.1
django-js-asset==0.1.1
django-mptt==0.7.4
django-sekizai==0.10.0
django-smtp-ssl==1.0
django-tagging==0.4
django-treebeard==3.0
djangocms-admin-style==1.2.7
docopt==0.6.2
docutils==0.14
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.1.2
greenlet==0.4.10
html5lib==1.0b8
httplib2==0.9.2
idna==2.6
imagesize==0.7.1
ipaddress==1.0.18
ipykernel==4.6.1
ipyparallel==6.0.2
ipython-genutils==0.2.0
ipython==5.4.1
ipywidgets==7.0.0
jdcal==1.3
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
kazoo==2.2.1
matplotlib==1.3.1
mistune==0.7.4
mots-vides==2015.5.11
mysql==0.0.1
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.0.0
numpy==1.8.0
odfpy==1.3.5
openpyxl==2.4.9
pandas==0.20.3
pandoc==1.0.0a3
pandocfilters==1.4.2
paramiko==2.2.1
passlib==1.6.5
path.py==8.1.2
pathlib2==2.3.0
pefile==2017.9.3
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
ply==3.10
prompt-toolkit==1.0.15
psutil==3.3.0
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.4.7
pydub==0.20.0
pyflakes==1.6.0
pykafka==2.5.0
pyparsing==2.2.0
pysmi==0.1.3
pysnmp==4.3.9
pysolr==3.6.0
pyte==0.5.2
python-dateutil==2.6.1
python-nmap==0.6.1
python-ptrace==0.9.3
pytz==2017.2
pyvim==0.0.21
pyvmomi==6.5.0.2017.5.post1
pyzmq==16.0.2
qtconsole==4.3.1
redis==2.9.1
requests==2.18.4
scandir==1.5
sh==1.12.14
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sorl-thumbnail==12.3
sphinxcontrib-websupport==1.0.1
tablib==0.12.1
tabulate==0.7.7
terminado==0.6
testpath==0.3.1
tornado==4.3
traitlets==4.3.2
typing==3.6.2
uWSGI==2.0.14
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.2
xlrd==1.1.0
xlwt==1.3.0
